refactor: 사용자 조회 메서드 타입 수정

+++ src/main/java/com/aiary/aiary/domain/diary/service/DiaryService.java
@@ -3,12 +3,14 @@ package com.aiary.aiary.domain.diary.service;
 import com.aiary.aiary.domain.diary.dto.mapper.DiaryMapper;
 import com.aiary.aiary.domain.diary.dto.request.DiaryCreateRequest;
 import com.aiary.aiary.domain.diary.entity.Diary;
+import com.aiary.aiary.domain.diary.exception.DiaryNotFoundException;
 import com.aiary.aiary.domain.diary.repository.DiaryRepository;
 import com.aiary.aiary.domain.user.entity.User;
 import com.aiary.aiary.domain.user.repository.UserRepository;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
@@ -19,7 +21,7 @@ public class DiaryService {
     private final DiaryMapper diaryMapper;
 
     public void createDiary(DiaryCreateRequest diaryCreateRequest){
-        User findUser = userRepository.findById((long) diaryCreateRequest.getUserId()).orElseThrow(IllegalAccessError::new);
+        User findUser = userRepository.findById(diaryCreateRequest.getUserId()).orElseThrow(IllegalAccessError::new);
 
         Diary newDiary = diaryMapper.mapCreateRequestToEntity(diaryCreateRequest, findUser);
         diaryRepository.save(newDiary);
